Starting shared utils for scheduled messages. Replaced the contextToFixDatePicker where used
