+++ app/src/main/java/xyz/stream/messenger/fragment/ScheduledMessagesFragment.kt
@@ -82,34 +82,6 @@ class ScheduledMessagesFragment : Fragment(), ScheduledMessageClickListener {
         }
     }
 
-    // samsung messed up the date picker in some languages on Lollipop 5.0 and 5.1. Ugh.
-    // fixes this issue: http://stackoverflow.com/a/34853067
-    private val contextToFixDatePickerCrash: ContextWrapper
-        get() = object : ContextWrapper(fragmentActivity!!) {
-            private var wrappedResources: Resources? = null
-            override fun getResources(): Resources {
-                val r = super.getResources()
-                if (wrappedResources == null) {
-                    wrappedResources = object : Resources(r.assets, r.displayMetrics, r.configuration) {
-                        @Throws(Resources.NotFoundException::class)
-                        override fun getString(id: Int, vararg formatArgs: Any): String {
-                            return try {
-                                super.getString(id, *formatArgs)
-                            } catch (ifce: IllegalFormatConversionException) {
-                                Log.e("DatePickerDialogFix", "IllegalFormatConversionException Fixed!", ifce)
-                                var template = super.getString(id)
-                                template = template.replace(("%" + ifce.conversion).toRegex(), "%s")
-                                String.format(configuration.locale, template, *formatArgs)
-                            }
-
-                        }
-                    }
-                }
-
-                return wrappedResources!!
-            }
-        }
-
     override fun onCreateView(inflater: LayoutInflater, parent: ViewGroup?, savedInstanceState: Bundle?): View? =
             inflater.inflate(R.layout.fragment_schedule_messages, parent, false)
