+++ shared/src/main/java/xyz/stream/messenger/shared/util/ScheduledMessageUtils.kt
@@ -1,0 +1,39 @@
+package xyz.stream.messenger.shared.util
+
+import android.app.Activity
+import android.content.ContextWrapper
+import android.content.res.Resources
+import android.util.Log
+import java.util.*
+
+object ScheduledMessageUtils {
+
+    // samsung messed up the date picker in some languages on Lollipop 5.0 and 5.1. Ugh.
+    // fixes this issue: http://stackoverflow.com/a/34853067
+    fun getContextToFixDatePickerCrash(activity: Activity): ContextWrapper {
+        return object : ContextWrapper(activity!!) {
+            private var wrappedResources: Resources? = null
+            override fun getResources(): Resources {
+                val r = super.getResources()
+                if (wrappedResources == null) {
+                    wrappedResources = object : Resources(r.assets, r.displayMetrics, r.configuration) {
+                        @Throws(Resources.NotFoundException::class)
+                        override fun getString(id: Int, vararg formatArgs: Any): String {
+                            return try {
+                                super.getString(id, *formatArgs)
+                            } catch (ifce: IllegalFormatConversionException) {
+                                Log.e("DatePickerDialogFix", "IllegalFormatConversionException Fixed!", ifce)
+                                var template = super.getString(id)
+                                template = template.replace(("%" + ifce.conversion).toRegex(), "%s")
+                                String.format(configuration.locale, template, *formatArgs)
+                            }
+
+                        }
+                    }
+                }
+
+                return wrappedResources!!
+            }
+        }
+    }
+}
